rename : 패키지 이름 변경 fp -> fastcampus
