+++ src/main/java/com/fastcampus/persistence/CartRepository.java
@@ -1,8 +1,7 @@
-package com.fp.persistence;
+package com.fastcampus.persistence;
 
-import com.fp.domain.Cart;
+import com.fastcampus.domain.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface CartRepository extends JpaRepository<Cart, Long> {
 }
