+++ src/main/java/com/fastcampus/persistence/MemberRepository.java
@@ -1,12 +1,13 @@
-package com.fp.persistence;
+package com.fastcampus.persistence;
 
-import com.fp.domain.Member;
+import com.fastcampus.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
+
     Optional<Member> findByEmail(String email);
+
 }
