+++ src/main/java/com/fastcampus/web/dto/MemberDto.java
@@ -1,10 +1,8 @@
-package com.fp.web.dto;
+package com.fastcampus.web.dto;
 
 import lombok.Data;
 
 @Data
 public class MemberDto {
-
-
 
 }
