+++ src/test/java/com/fastcampus/FinancialProductRecommendationServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fp;
+package com.fastcampus;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
